backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-docx>=1.1.0
mammoth>=1.6.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
Pillow>=10.0.0
aiofiles>=23.2.0
pydantic>=2.5.0
